requirements.txt
flask>=3.0.0
python-dotenv>=1.0.1
psycopg[binary]>=3.1.8
werkzeug>=3.0.0
pillow>=10.0.0
google-generativeai>=0.5.0
